ディレクトリ構成修正,scssファイルをコーポネント毎に作成,Type.tsx削除
