+++ src/compornents/home/TodoList.tsx
@@ -4,8 +4,8 @@ import React, {
   Dispatch,
   SetStateAction,
 } from "react";
-import styles from "./CssModules.module.scss";
-import { Todo } from "./Type";
+import styles from "./todolist.module.scss";
+import { Todo } from "../../pages/Home";
 import { RadioBtn } from "./RadioBtn";
 import { TodoInput } from "./TodoInput";
 import { List } from "./List";
@@ -31,14 +31,17 @@ export const TextContext = createContext(
   }
 );
 
+export const URGENT = "急ぎ";
+export const MORNING = "午前";
+export const AFTERNOON = "午後";
+
+export const WORK_ON_PROGRESS = "作業中";
+export const DONE = "完了";
+
 export const TodoList = () => {
   const [radioCategoly, setRadioCategoly] = useState<string>("午前");
   const [todos, setTodos] = useState<Todo[]>([]);
   const [text, setText] = useState<string>("");
-
-  const URGENT = "急ぎ";
-  const MORNING = "午前";
-  const AFTERNOON = "午後";
 
   return (
     <div className={styles.todo_list}>
@@ -48,9 +51,9 @@ export const TodoList = () => {
       <RadioContext.Provider value={{ radioCategoly, setRadioCategoly }}>
         <TodosContext.Provider value={{ todos, setTodos }}>
           <TextContext.Provider value={{ text, setText }}>
-            <TodoInput />
+            <TodoInput WORK_ON_PROGRESS={WORK_ON_PROGRESS} />
           </TextContext.Provider>
-          <RadioBtn />
+          <RadioBtn URGENT={URGENT} MORNING={MORNING} AFTERNOON={AFTERNOON} />
           <List categoly={URGENT} />
           <List categoly={MORNING} />
           <List categoly={AFTERNOON} />
