+++ src/compornents/home/todolist.module.scss
@@ -1,34 +1,3 @@
-.container {
-  display: flex;
-  box-sizing: border-box;
-}
-
-.calendar {
-  width: 70%;
-}
-
-.input_text_area {
-  display: flex;
-  .t_p {
-    margin-left: 20px;
-  }
-  .t_input {
-    margin-left: 20px;
-    margin-top: 18px;
-    height: 15px;
-  }
-  .t_button {
-    margin-left: 20px;
-    margin-top: 15px;
-    height: 25px;
-  }
-}
-
-.todo_list {
-  margin-left: auto;
-  margin-right: auto;
-}
-
 .title {
   text-align: center;
   font-size: 20px;
@@ -61,6 +30,7 @@
 .list_container {
   margin-bottom: 30px;
   background-color: #f5f5f5;
+  height: 150px;
 }
 
 .subtitle {
@@ -84,8 +54,4 @@
     margin: 0 5px 0 0;
     font-size: 7px;
   }
-}
-
-.carried_over {
-  margin-top: 70px;
 }
